requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==3.2.21
django-cors-headers==4.1.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==1.2.0
gunicorn==21.2.0
ipython==7.34.0
jedi==0.19.0
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
